IPE-468 transaction evaluation integration test completed

+++ src/main/java/it/gov/pagopa/initiative/statistics/dto/InitiativeStatisticsDTO.java
@@ -17,7 +17,7 @@ import java.time.OffsetDateTime;
 public class InitiativeStatisticsDTO {
 
   private OffsetDateTime lastUpdatedDateTime;
-  private Integer onboardedCitizenCount;
+  private Long onboardedCitizenCount;
   private String accruedRewards;
 
 }
